+++ config.go
@@ -3,8 +3,9 @@ package main
 import (
 	"flag"
 	"fmt"
-	"strings"
 	"os"
+	"runtime"
+	"strings"
 )
 
 const (
@@ -72,8 +73,11 @@ func (config *SwampConfig) SetupFlags() {
 	flag.StringVar(&config.tokenSerialNumber, "mfa-device", config.tokenSerialNumber, "MFA device arn")
 	flag.BoolVar(&config.useInstanceProfile, "instance", config.useInstanceProfile, "Use instance profile")
 	flag.BoolVar(&config.renew, "renew", config.renew, "renew token every duration/2")
-	flag.BoolVar(&config.exportProfile, "export-profile", config.exportProfile, "set AWS_PROFILE in environment")
-	flag.StringVar(&config.exportFile, "export-file", config.exportFile, "File to write AWS_PROFILE to, defaults to '/tmp/current_swamp_profile'")
+	if runtime.GOOS == "linux" || runtime.GOOS == "darwin" {
+		// platform specific flags
+		flag.BoolVar(&config.exportProfile, "export-profile", config.exportProfile, "set AWS_PROFILE in environment")
+		flag.StringVar(&config.exportFile, "export-file", config.exportFile, "File to write AWS_PROFILE to, defaults to '/tmp/current_swamp_profile'")
+	}
 	flag.Usage = flagUsage
 }