+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * obey environment variable `AWS_SHARED_CREDENTIALS_FILE`
 * hide flags `-export-profile` and `-export-file` on windows
+* `-target-profile` defaults to `swamp`
 
 ## swamp v0.6